v1.3: fix truncated kernel grids, speed up the per-template pipeline, drop non-CUDA code
Correctness
- resampling/padding/modulation kernels were launched with nsamples/blocksize
  blocks (integer division); for lengths that are not multiples of 384/512 the
  tail of the resampled series was never written and kept the previous
  template's FFT output -> bogus low-frequency candidates. Grids are rounded up
  and bounds-checked, the mean reduction handles partial blocks.
- verified against the upstream CPU application (FFTW) on synthetic strong and
  weak signals: 100/100 candidates agree, powers within 1e-5 relative;
  Windows and Linux builds give bit-identical candidate lists.

Performance (same numerical results)
- block-parallel modulated-length search instead of a single-thread walk
- resampled length and mean consumed on the device: no device->host round
  trips inside run_resampling
- harmonic summing: device/pinned buffers and streams allocated once,
  asynchronous threshold upload and read-back, two sync points per template
- ~24 % less time per template on an RTX 5070 Ti (4M samples, 512 templates)

Cleanup
- removed upstream OpenCL/Metal/CPU backends, non-CUDA makefiles, cuPrintf,
  build.sh and patches/; demod_binary.c is CUDA-only
- cores-per-SM table shared between banner and device selection
- scripts locate the repository from their own path; obsolete CUDA 11.8
  legacy scripts removed; test runner takes the executable as argument;
  make_synthetic gained --amp for weak-signal data
- README/RELEASES/scripts docs updated, banner bumped to v1.3

+++ src/erp_utilities.cpp
@@ -122,25 +122,25 @@ void logMessage(const ERP_LOGLEVEL logLevel, const bool showLevel, const char *m
   switch (logLevel) {
     case error:
       output = stderr;
-      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][ERROR] ", timeBuffer, pid);
+      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][ERROR] ", timeBuffer, (long long)pid);
       break;
     case warn:
       output = stderr;
-      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][WARN ] ", timeBuffer, pid);
+      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][WARN ] ", timeBuffer, (long long)pid);
       break;
     case info:
       // we want these to be logged by BOINC
       output = stderr;
-      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][INFO ] ", timeBuffer, pid);
+      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][INFO ] ", timeBuffer, (long long)pid);
       break;
     case debug:
       // we want these to be logged by BOINC
       output = stderr;
-      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][DEBUG] ", timeBuffer, pid);
+      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][DEBUG] ", timeBuffer, (long long)pid);
       break;
     default:
       output = stderr;
-      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][UNKWN] ", timeBuffer, pid);
+      snprintf(levelBuffer, LEVEL_BUFFER_SIZE, "[%s][%" FORMAT "][UNKWN] ", timeBuffer, (long long)pid);
       break;
   }
